[FIX] account_invoice_margin: avoid traceback in invoice
When creating an invoice, associating a product
(the uom is associated), deleting the product from the
line and saving, an error is obtained:

  File
"/home/odoo/src/user/margin-analysis/account_invoice_margin/models/account_invoice.py",
line 108, in _compute_purchase_price
    purchase_price = line.product_id.uom_id._compute_price(
  File "/home/odoo/src/odoo/addons/uom/models/uom_uom.py", line 247, in
_compute_price
    self.ensure_one()
  File "/home/odoo/src/odoo/odoo/models.py", line 5243, in ensure_one
    raise ValueError("Expected singleton: %s" % self)

This is because it attempts to take the uom of the
product to calculate the purchase price and the
product is empty.

The product is validated before performing the
calculation, to avoid the error.